+++ core/messages/de/luya.php
@@ -1,5 +1,11 @@
 <?php
 
 return [
-    'The string must have a minimum length of {length} characters.' => 'Die Zeichenkette muss mindistens über {length} Zeichen verfügen.',
+    '{attribute} must have a minimum length of {length} characters.' => '{attribute} muss mindestens über {length} Zeichen verfügen.',
+    '{attribute} must include at least one special character.' => '{attribute} muss mindestens ein Spezial-Zeichen enthalten.',
+    '{attribute} must include at least one digit.' => '{attribute} muss mindestens eine Zahl enthalten.',
+    '{attribute} must include at least one letter.' => '{attribute} muss mindestens einen Buchstaben enthalten.',
+    '{attribute} must include at least one uppercase letter.' => '{attribute} muss mindestens einen Grossbuchstaben enthalten.',
+    '{attribute} must include at least one lowercase letter.' => '{attribute} muss mindestens einen Kleinbuchstaben enthalten.',
+    '{attribute} must be a float or numeric value.' => '{attribute} muss ein Numerischer Wert sein.'
 ];